Limit local_ratelimit source files to one logger ID
